supir_rest_api/requirements.txt
fastapi
uvicorn
python-multipart
gradiotorch
torchvision
python-dotenv
